skills/automation_lab/tool_wiki_video/requirements.txt
gTTS
moviepy
requests
wikipedia
svglib
reportlab
opencv-python
Pillow
imageio
